+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- AsyncFileWriter utility
 - `Tools/Editor/Open Application.persistentDataPath` menu option
 
 ## [1.3.0] - 2023-10-12